composetest/auth/requirements.txt
flask
redis
PyJWT